+++ src/app/products/components/product-list/product-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ProductModel } from '../../models/product-model';
 import { ProductService } from '../../services/product.service';
 import { CartService } from 'src/app/cart/services/cart.service';
+import { OrderByPipe } from 'src/app/shared/pipes/order-by.pipe';
 
 @Component({
   selector: 'app-product-list',
@@ -10,7 +11,11 @@ import { CartService } from 'src/app/cart/services/cart.service';
 })
 export class ProductListComponent {
 
+  selectedSortParam!: string;
+  isAsc: boolean = true;
+
   constructor(
+    public orderByPipe: OrderByPipe,
     public productService: ProductService,
     private cartService: CartService
   ) {}
@@ -21,4 +26,9 @@ export class ProductListComponent {
       console.log("You have bought " + product.name + ". You have to pay " + product.price)
     }
   }
+
+  onSortParamChange(event: any) {
+    this.selectedSortParam = event.target.value;
+  }
+
 }