+++ src/app/products/products.module.ts
@@ -3,6 +3,7 @@ import { NgModule } from "@angular/core";
 import { ProductComponent } from "./components/product/product.component";
 import { ProductListComponent } from "./components/product-list/product-list.component";
 import { SharedModule } from "../shared/shared.module";
+import { OrderByPipe } from "../shared/pipes/order-by.pipe";
 
 @NgModule({
     imports: [
@@ -15,6 +16,9 @@ import { SharedModule } from "../shared/shared.module";
     ],
     exports: [
         ProductListComponent
+    ],
+    providers: [
+        OrderByPipe
     ]
 })
 export class ProductsModule {}